+++ zerver/tests/test_markdown.py
@@ -3181,12 +3181,16 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
             f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/">#{denmark.name} &gt; <em>{Message.EMPTY_TOPIC_FALLBACK_NAME}</em></a></p>',
         )
 
-    def test_topic_single_containing_message(self) -> None:
+    def test_topic_single_containing_messages(self) -> None:
         denmark = get_stream("Denmark", get_realm("zulip"))
         sender_user_profile = self.example_user("othello")
-        first_message_id = self.send_stream_message(
+        self.send_stream_message(
             sender_user_profile, "Denmark", topic_name="some topic", content="test"
         )
+        latest_message_id = self.send_stream_message(
+            sender_user_profile, "Denmark", topic_name="some topic", content="test 2"
+        )
+
         msg = Message(
             sender=sender_user_profile,
             sending_client=get_client("test"),
@@ -3195,7 +3199,7 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
         content = "#**Denmark>some topic**"
         self.assertEqual(
             render_message_markdown(msg, content).rendered_content,
-            f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{first_message_id}">#{denmark.name} &gt; some topic</a></p>',
+            f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{latest_message_id}">#{denmark.name} &gt; some topic</a></p>',
         )
 
     def test_topic_atomic_string(self) -> None:
@@ -3227,8 +3231,11 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
         denmark = get_stream("Denmark", get_realm("zulip"))
         scotland = get_stream("Scotland", get_realm("zulip"))
         sender_user_profile = self.example_user("othello")
-        first_message_id = self.send_stream_message(
+        self.send_stream_message(
             sender_user_profile, "Denmark", topic_name="some topic", content="test"
+        )
+        latest_message_id = self.send_stream_message(
+            sender_user_profile, "Denmark", topic_name="some topic", content="test 2"
         )
         msg = Message(
             sender=sender_user_profile,
@@ -3240,7 +3247,7 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
             render_message_markdown(msg, content).rendered_content,
             "<p>This has two links: "
             f'<a class="stream-topic" data-stream-id="{denmark.id}" '
-            f'href="/#narrow/channel/{denmark.id}-{denmark.name}/topic/some.20topic/with/{first_message_id}">'
+            f'href="/#narrow/channel/{denmark.id}-{denmark.name}/topic/some.20topic/with/{latest_message_id}">'
             f"#{denmark.name} &gt; some topic</a>"
             " and "
             f'<a class="stream-topic" data-stream-id="{scotland.id}" '
@@ -3253,8 +3260,11 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
         realm = get_realm("zulip")
         denmark = get_stream("Denmark", get_realm("zulip"))
         sender_user_profile = self.example_user("othello")
-        first_message_id = self.send_stream_message(
+        self.send_stream_message(
             sender_user_profile, "Denmark", topic_name="some topic", content="test"
+        )
+        latest_message_id = self.send_stream_message(
+            sender_user_profile, "Denmark", topic_name="some topic", content="test 2"
         )
 
         msg = Message(
@@ -3275,7 +3285,7 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
         ):
             self.assertEqual(
                 render_message_markdown(msg, content, mention_data=mention_data).rendered_content,
-                f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{first_message_id}">#{denmark.name} &gt; some topic</a></p>',
+                f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{latest_message_id}">#{denmark.name} &gt; some topic</a></p>',
             )
 
         # test topic linked doesn't have any message in it in case
@@ -3291,7 +3301,7 @@ class MarkdownStreamTopicMentionTests(ZulipTestCase):
         content = "#**Denmark>some topic**"
         self.assertEqual(
             markdown_convert_wrapper(content),
-            f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{first_message_id}">#{denmark.name} &gt; some topic</a></p>',
+            f'<p><a class="stream-topic" data-stream-id="{denmark.id}" href="/#narrow/channel/{denmark.id}-Denmark/topic/some.20topic/with/{latest_message_id}">#{denmark.name} &gt; some topic</a></p>',
         )
 
         # test topic links for channel with protected history